hotfix: added snackbar in app.vue refs #201
