+++ src/App.vue
@@ -18,18 +18,63 @@
         />
       </div>
     </transition>
+    <SnackBar :showSnackBar="showPrivacySnackBar">
+      <div
+        class="flex items-center justify-between text-white w-full h-12 tg-caption-mobile leading-4 p-4"
+      >
+        <p>
+          Gotta agree to
+          <a
+            class="underline cursor-pointer"
+            target="_blank"
+            href="https://hub.paulchrisluke.com/-temporary-slug-7a760197-2d5d-4314-876c-ade5923d6dd8"
+          >
+            terms and conditions
+          </a>
+        </p>
+        <p
+          class="text-button uppercase cursor-pointer"
+          @click="setSnackBarCookie"
+        >
+          Agree
+        </p>
+      </div>
+    </SnackBar>
   </div>
 </template>
 
 <script>
 import BaseOverlaySuccess from '@/components/BaseOverlaySuccess.vue';
+import SnackBar from '@/components/SnackBar.vue';
+import cookie from '@/utils/cookie';
 
 export default {
   name: 'App',
-  components: { BaseOverlaySuccess },
+  components: { BaseOverlaySuccess, SnackBar },
   computed: {
     overlayModel() {
       return this.$store.getters['overlay/model'];
+    }
+  },
+  data() {
+    return {
+      showPrivacySnackBar: true
+    };
+  },
+  beforeMount() {
+    let showSnackBar = cookie.getCookie('privacy-snackbar');
+    //if bannerCookie === 1 that means user has seen the banner and dismissed it
+    if (showSnackBar == 1) {
+      this.showPrivacySnackBar = false;
+    } else {
+      this.showPrivacySnackBar = true;
+    }
+  },
+  methods: {
+    setSnackBarCookie() {
+      this.showPrivacySnackBar = false;
+      //set cookie with name 'snackbar'. Set value to 1 which means true. Set expiration to 7 days.
+      cookie.setCookie('privacy-snackbar', 1, 7);
     }
   }
 };
