+++ src/views/Home.vue
@@ -57,36 +57,11 @@
       </ul>
     </section>
     <CreditFooter />
-    <SnackBar :showSnackBar="showPrivacySnackBar">
-      <div
-        class="flex items-center justify-between text-white w-full h-12 tg-caption-mobile leading-4 p-4"
-      >
-        <p>
-          Gotta agree to
-          <a
-            class="underline cursor-pointer"
-            target="_blank"
-            href="https://hub.paulchrisluke.com/-temporary-slug-7a760197-2d5d-4314-876c-ade5923d6dd8"
-          >
-            terms and conditions
-          </a>
-        </p>
-        <p
-          class="text-button uppercase cursor-pointer"
-          @click="setSnackBarCookie"
-        >
-          Agree
-        </p>
-      </div>
-    </SnackBar>
   </div>
 </template>
 
 <script>
-import cookie from '@/utils/cookie';
-
 import CreditFooter from '@/components/CreditFooter.vue';
-import SnackBar from '@/components/SnackBar.vue';
 import TenantCard from '@/components/tenant/TenantCard.vue';
 import { mapGetters, mapActions } from 'vuex';
 
@@ -94,36 +69,17 @@ export default {
   name: 'Home',
   components: {
     CreditFooter,
-    SnackBar,
     TenantCard
-  },
-  data() {
-    return {
-      showPrivacySnackBar: true
-    };
   },
   created() {
     this.fetchHomeData();
   },
-  beforeMount() {
-    let showSnackBar = cookie.getCookie('privacy-snackbar');
-    //if bannerCookie === 1 that means user has seen the banner and dismissed it
-    if (showSnackBar == 1) {
-      this.showPrivacySnackBar = false;
-    } else {
-      this.showPrivacySnackBar = true;
-    }
-  },
+
   mounted() {
     this.fetchTenants();
   },
   methods: {
-    ...mapActions('home', ['fetchHomeData', 'fetchTenants']),
-    setSnackBarCookie() {
-      this.showPrivacySnackBar = false;
-      //set cookie with name 'snackbar'. Set value to 1 which means true. Set expiration to 7 days.
-      cookie.setCookie('privacy-snackbar', 1, 7);
-    }
+    ...mapActions('home', ['fetchHomeData', 'fetchTenants'])
   },
   computed: {
     ...mapGetters({
